refactor: inline custom model path and fix test cleanup
