--- app/src/constants/modelPaths.ts
@@ -1,5 +1,0 @@
-import * as FileSystem from 'expo-file-system';
-
-// Path where the downloaded centroid model JSON is stored
-export const CUSTOM_GESTURE_MODEL_PATH =
-  FileSystem.documentDirectory + 'custom_model.json';
