+++ app/src/screens/AdminScreen.tsx
@@ -21,7 +21,7 @@ import * as FileSystem from 'expo-file-system';
 import { API_URL } from '../constants';
 import { database } from '../../db';
 import { useServices } from '../context/ServicesContext';
-import { CUSTOM_GESTURE_MODEL_PATH } from '../constants/modelPaths';
+import { CUSTOM_GESTURE_MODEL_PATH } from '../constants';
 import { CUSTOM_AUDIO_DIR, getCustomAudioPath } from '../constants/audioPaths';
 import { Symbol as DBSymbol } from '../../db/models';
 import { COLORS, SPACING, RADIUS } from '../constants/ui';
